requirements.txt
groq==0.4.1
httpx==0.27.2
python-dotenv==1.0.0
flask==3.0.0